+++ src/main/java/org/jolokia/docker/maven/RemoveMojo.java
@@ -38,18 +38,19 @@ import org.jolokia.docker.maven.config.*;
 
 public class RemoveMojo extends AbstractDockerMojo {
 
-    // Whether all configured images should be removed
     /**
+     * Should all configured images should be removed?
+     * 
      * @parameter property = "docker.removeAll" default-value = "false"
      */
     private boolean removeAll;
-
+    
     @Override
     protected void executeInternal(DockerAccess dockerAccess) throws DockerAccessException {
         for (ImageConfiguration image : getImages()) {
             String name = image.getName();
             if (removeAll || image.isDataImage()) {
-                if (dockerAccess.hasImage(name)) {
+                if (queryService.hasImage(name)) {
                     if (dockerAccess.removeImage(name,true)) {
                         log.info(image.getDescription() + ": Remove");
                     }