+++ src/main/java/org/jolokia/docker/maven/StartMojo.java
@@ -20,6 +20,7 @@ import org.jolokia.docker.maven.access.log.LogCallback;
 import org.jolokia.docker.maven.access.log.LogGetHandle;
 import org.jolokia.docker.maven.config.*;
 import org.jolokia.docker.maven.log.LogDispatcher;
+import org.jolokia.docker.maven.service.QueryService;
 import org.jolokia.docker.maven.service.RunService;
 import org.jolokia.docker.maven.util.*;
 
@@ -47,8 +48,9 @@ public class StartMojo extends AbstractDockerMojo {
     protected RunService runService;
 
     @Override
-    protected void initLog(Logger log) {
-        runService.initLog(log);
+    protected void initializeServices(DockerAccess access, QueryService queryService, Logger log) {
+        super.initializeServices(access, queryService, log);
+        runService.initialize(access, queryService, log);
     }
 
     /**
@@ -56,26 +58,24 @@ public class StartMojo extends AbstractDockerMojo {
      */
     @Override
     public synchronized void executeInternal(final DockerAccess dockerAccess) throws DockerAccessException, MojoExecutionException {
-
         getPluginContext().put(CONTEXT_KEY_START_CALLED, true);
 
         LogDispatcher dispatcher = getLogDispatcher(dockerAccess);
         try {
-            for (StartOrderResolver.Resolvable resolvable : runService.getImagesConfigsInOrder(getImages())) {
+            for (StartOrderResolver.Resolvable resolvable : runService.getImagesConfigsInOrder(queryService, getImages())) {
                 final ImageConfiguration imageConfig = (ImageConfiguration) resolvable;
 
                 // Still to check: How to work with linking, volumes, etc ....
                 //String imageName = new ImageName(imageConfig.getName()).getFullNameWithTag(registry);
 
                 String imageName = imageConfig.getName();
-
                 checkImageWithAutoPull(dockerAccess, imageName,
                                        getConfiguredRegistry(imageConfig),imageConfig.getBuildConfiguration() == null);
 
                 RunImageConfiguration runConfig = imageConfig.getRunConfiguration();
                 PortMapping portMapping = runService.getPortMapping(runConfig, project.getProperties());
 
-                String containerId = runService.createAndStartContainer(dockerAccess, imageConfig, portMapping, project.getProperties());
+                String containerId = runService.createAndStartContainer(imageConfig, portMapping, project.getProperties());
 
                 if (showLogs(imageConfig)) {
                     dispatcher.trackContainerLog(containerId, getContainerLogSpec(containerId, imageConfig));
@@ -88,14 +88,14 @@ public class StartMojo extends AbstractDockerMojo {
                 waitIfRequested(dockerAccess,imageConfig, project.getProperties(), containerId);
             }
             if (follow) {
-                runService.addShutdownHookForStoppingContainers(dockerAccess,keepContainer,removeVolumes);
+                runService.addShutdownHookForStoppingContainers(keepContainer,removeVolumes);
                 wait();
             }
         } catch (InterruptedException e) {
             log.warn("Interrupted");
         }
     }
-
+    
     private void updateDynamicPortProperties(DockerAccess docker, String containerId, RunImageConfiguration runConfig, PortMapping mappedPorts, Properties properties) throws DockerAccessException, MojoExecutionException {
         if (mappedPorts.containsDynamicPorts()) {
             mappedPorts.updateVariablesWithDynamicPorts(docker.queryContainerPortMapping(containerId));