+++ src/main/java/org/jolokia/docker/maven/StopMojo.java
@@ -5,6 +5,8 @@ import org.jolokia.docker.maven.access.DockerAccess;
 import org.jolokia.docker.maven.access.DockerAccessException;
 import org.jolokia.docker.maven.config.ImageConfiguration;
 import org.jolokia.docker.maven.log.LogDispatcher;
+import org.jolokia.docker.maven.model.Container;
+import org.jolokia.docker.maven.service.QueryService;
 import org.jolokia.docker.maven.service.RunService;
 import org.jolokia.docker.maven.util.Logger;
 
@@ -34,10 +36,11 @@ public class StopMojo extends AbstractDockerMojo {
 
     /** @component */
     private RunService runService;
-
+    
     @Override
-    protected void initLog(Logger log) {
-        runService.initLog(log);
+    protected void initializeServices(DockerAccess access, QueryService queryService, Logger log) {
+        super.initializeServices(access, queryService, log);
+        runService.initialize(access, queryService, log);
     }
 
     @Override
@@ -49,12 +52,12 @@ public class StopMojo extends AbstractDockerMojo {
                 // Called directly ....
                 for (ImageConfiguration image : getImages()) {
                     String imageName = image.getName();
-                    for (String container : access.getContainersForImage(imageName)) {
-                        runService.stopContainer(access, image, container, keepContainer, removeVolumes);
+                    for (Container container : queryService.getContainersForImage(imageName)) {
+                        runService.stopContainer(image, container.getId(), keepContainer, removeVolumes);
                     }
                 }
             } else {
-                runService.stopStartedContainers(access, keepContainer, removeVolumes);
+                runService.stopStartedContainers(keepContainer, removeVolumes);
             }
         }